have sprites with float for layer

+++ include/mpUtils/Graphics/Rendering2D/Renderer2D.h
@@ -90,7 +90,7 @@ public:
      *          ordering within layers is done with floating point numbers
      * @param color add tint color
      */
-    void addSprite(const Sprite2D& sprite, const glm::mat4& transform=glm::mat4(1.0f), int layer=0, const glm::vec4& color=glm::vec4(1.0f));
+    void addSprite(const Sprite2D& sprite, const glm::mat4& transform=glm::mat4(1.0f), float layer=0, const glm::vec4& color=glm::vec4(1.0f));
 
     /**
      * @brief render everything queued up for this frame then clear queue
